perf: 优化代码结构
