+++ apps/ops/ansible/__init__.py
@@ -5,3 +5,4 @@ from .inventory import *
 from .runners import *
 from .exceptions import *
 from .runner import *
+from .interface import *
