+++ apps/ops/ansible/runner.py
@@ -4,31 +4,12 @@ import uuid
 
 from django.conf import settings
 from django.utils._os import safe_join
-from django.utils.functional import LazyObject
-
+from . import interface
 from .callback import DefaultCallback
 from .exception import CommandInBlackListException
-from .runners import AnsibleReceptorRunner, AnsibleNativeRunner, RunnerInterface
 from ..utils import get_ansible_log_verbosity
 
-__all__ = ['AdHocRunner', 'PlaybookRunner', 'SuperPlaybookRunner', 'UploadFileRunner', 'interface']
-
-
-class _LazyRunnerInterface(LazyObject):
-
-    def _setup(self):
-        self._wrapped = self.make_interface()
-
-    @staticmethod
-    def make_interface():
-        runner_type = AnsibleReceptorRunner \
-            if settings.ANSIBLE_RECEPTOR_ENABLED else AnsibleNativeRunner
-        gateway_host = settings.ANSIBLE_RECEPTOR_GATEWAY_PROXY_HOST \
-            if settings.ANSIBLE_RECEPTOR_GATEWAY_PROXY_HOST else '127.0.0.1'
-        return RunnerInterface(runner_type=runner_type, gateway_proxy_host=gateway_host)
-
-
-interface = _LazyRunnerInterface()
+__all__ = ['AdHocRunner', 'PlaybookRunner', 'SuperPlaybookRunner', 'UploadFileRunner']
 
 
 class AdHocRunner:
