+++ apps/ops/ansible/runners/__init__.py
@@ -1,4 +1,3 @@
-from .interface import *
 from .base import *
 from .native import *
 from .receptor import *
